CSP-1473: Entension of the PostPermissions endpoint to handle notifications.

+++ src/EmployerPR/SFA.DAS.EmployerPR.UnitTests/Application/Providers/Queries/GetPermissionsHas/GetPermissionsHasQueryHandlerTests.cs
@@ -21,7 +21,7 @@ public class GetPermissionsQueryHandlerTests
         providerRelationshipsApiRestClient.Setup(x =>
             x.GetPermissions(
                 It.IsAny<long>(),
-                It.IsAny<int>(),
+                It.IsAny<long>(),
                 It.IsAny<CancellationToken>()
             )
         )
